HW3
Asks user for two inputs and performs choice of math operation. 1st time through, it doesn't include the quit option. Anytime after, it does. To do this, I used variable i and incremented it after the first run through. So when i=0, it prints without quit option. When i>0, it prints with the quit option.